modified match parser

+++ src/main/java/org/all/info/dao/match/GameModeDAO.java
@@ -10,6 +10,6 @@ public interface GameModeDAO {
 
     public GameMode read(String name);
 
-    public GameMode read(Long id);
+    public GameMode read(Integer id);
 
 }
